+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/person/ui/overview/PersonOverviewFragment.kt
@@ -8,6 +8,7 @@ import androidx.recyclerview.widget.DefaultItemAnimator
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.covidcertificate.person.ui.overview.items.CertificatesItem
 import de.rki.coronawarnapp.databinding.PersonOverviewFragmentBinding
+import de.rki.coronawarnapp.util.DialogHelper
 import de.rki.coronawarnapp.util.di.AutoInject
 import de.rki.coronawarnapp.util.lists.decorations.TopBottomPaddingDecorator
 import de.rki.coronawarnapp.util.lists.diffutil.update
@@ -26,8 +27,47 @@ class PersonOverviewFragment : Fragment(R.layout.person_overview_fragment), Auto
     private val personOverviewAdapter = PersonOverviewAdapter()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        binding.bindRecycler()
-        binding.toolbar.setOnMenuItemClickListener {
+        binding.apply {
+            bindToolbar()
+            bindRecycler()
+        }
+        viewModel.personCertificates.observe(viewLifecycleOwner) { binding.bindViews(it) }
+        viewModel.events.observe(viewLifecycleOwner) { onNavEvent(it) }
+    }
+
+    private fun onNavEvent(event: PersonOverviewFragmentEvents) {
+        when (event) {
+            is OpenPersonDetailsFragment -> TODO()
+            is ShowDeleteDialog -> {
+                val dialog = DialogHelper.DialogInstance(
+                    context = requireContext(),
+                    title = R.string.test_certificate_delete_dialog_title,
+                    message = R.string.test_certificate_delete_dialog_body,
+                    positiveButton = R.string.test_certificate_delete_dialog_confirm_button,
+                    negativeButton = R.string.test_certificate_delete_dialog_cancel_button,
+                    cancelable = false,
+                    positiveButtonFunction = {
+                        viewModel.deleteTestCertificate(event.certificateId)
+                    }
+                )
+                DialogHelper.showDialog(dialog)
+            }
+            is ShowRefreshErrorDialog -> {
+                val dialog = DialogHelper.DialogInstance(
+                    context = requireContext(),
+                    title = R.string.test_certificate_refresh_dialog_title,
+                    message = event.error.localizedMessage ?: getString(R.string.errors_generic_headline),
+                    positiveButton = R.string.test_certificate_refresh_dialog_confirm_button,
+                    cancelable = false
+                )
+                DialogHelper.showDialog(dialog)
+            }
+            ScanQrCode -> TODO()
+        }
+    }
+
+    private fun PersonOverviewFragmentBinding.bindToolbar() {
+        toolbar.setOnMenuItemClickListener {
             when (it.itemId) {
                 R.id.menu_information -> doNavigate(
                     PersonOverviewFragmentDirections.actionPersonOverviewFragmentToVaccinationConsentFragment(false)
@@ -36,7 +76,6 @@ class PersonOverviewFragment : Fragment(R.layout.person_overview_fragment), Auto
                 else -> onOptionsItemSelected(it)
             }
         }
-        viewModel.personCertificates.observe(viewLifecycleOwner) { binding.bindViews(it) }
     }
 
     private fun PersonOverviewFragmentBinding.bindViews(persons: List<CertificatesItem>) {
