+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/test/ui/CertificatesFragment.kt
@@ -9,11 +9,9 @@ import androidx.recyclerview.widget.DefaultItemAnimator
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.covidcertificate.vaccination.ui.list.VaccinationListFragment
 import de.rki.coronawarnapp.databinding.FragmentCertificatesBinding
-import de.rki.coronawarnapp.util.DialogHelper
 import de.rki.coronawarnapp.util.di.AutoInject
 import de.rki.coronawarnapp.util.lists.decorations.TopBottomPaddingDecorator
 import de.rki.coronawarnapp.util.lists.diffutil.update
-import de.rki.coronawarnapp.util.ui.doNavigate
 import de.rki.coronawarnapp.util.ui.findNestedGraph
 import de.rki.coronawarnapp.util.ui.observe2
 import de.rki.coronawarnapp.util.ui.viewBinding
@@ -49,7 +47,7 @@ class CertificatesFragment : Fragment(R.layout.fragment_certificates), AutoInjec
                 )
                 is CertificatesFragmentEvents.OpenVaccinationRegistrationGraph -> {
                     findNestedGraph(R.id.vaccination_nav_graph).startDestination = R.id.vaccinationQrCodeScanFragment
-                    //doNavigate(CertificatesFragmentDirections.actionCertificatesFragmentToVaccinationNavGraph())
+                    // doNavigate(CertificatesFragmentDirections.actionCertificatesFragmentToVaccinationNavGraph())
                 }
                 is CertificatesFragmentEvents.GoToCovidCertificateDetailScreen -> {
                     //      doNavigate(
@@ -58,28 +56,8 @@ class CertificatesFragment : Fragment(R.layout.fragment_certificates), AutoInjec
                     //      )
                 }
                 is CertificatesFragmentEvents.ShowRefreshErrorCertificateDialog -> {
-                    val dialog = DialogHelper.DialogInstance(
-                        context = requireContext(),
-                        title = R.string.test_certificate_refresh_dialog_title,
-                        message = event.error.localizedMessage,
-                        positiveButton = R.string.test_certificate_refresh_dialog_confirm_button,
-                        cancelable = false
-                    )
-                    DialogHelper.showDialog(dialog)
                 }
                 is CertificatesFragmentEvents.ShowDeleteErrorCertificateDialog -> {
-                    val dialog = DialogHelper.DialogInstance(
-                        context = requireContext(),
-                        title = R.string.test_certificate_delete_dialog_title,
-                        message = R.string.test_certificate_delete_dialog_body,
-                        positiveButton = R.string.test_certificate_delete_dialog_confirm_button,
-                        negativeButton = R.string.test_certificate_delete_dialog_cancel_button,
-                        cancelable = false,
-                        positiveButtonFunction = {
-                            viewModel.deleteTestCertificate(event.identifier)
-                        }
-                    )
-                    DialogHelper.showDialog(dialog)
                 }
             }
         }
@@ -89,7 +67,7 @@ class CertificatesFragment : Fragment(R.layout.fragment_certificates), AutoInjec
         setOnMenuItemClickListener {
             when (it.itemId) {
                 R.id.menu_information -> {
-                    //doNavigate(CertificatesFragmentDirections.actionCertificatesFragmentToConsentFragment())
+                    // doNavigate(CertificatesFragmentDirections.actionCertificatesFragmentToConsentFragment())
                     true
                 }
                 else -> onOptionsItemSelected(it)
